+++ functions_visualization.py
